+++ pybricks/pupdevices/pb_module_pupdevices.c
@@ -23,7 +23,7 @@ static const mp_rom_map_elem_t pupdevices_globals_table[] = {
     { MP_ROM_QSTR(MP_QSTR_Light),               MP_ROM_PTR(&pb_type_pupdevices_Light)              },
     { MP_ROM_QSTR(MP_QSTR_PFMotor),             MP_ROM_PTR(&pb_type_pupdevices_PFMotor)            },
     #if PYBRICKS_PY_PUPDEVICES_DUPLO_TRAIN
-    { MP_ROM_QSTR(MP_QSTR_DuploTrain),          MP_ROM_PTR(&pb_type_remote)                        },
+    { MP_ROM_QSTR(MP_QSTR_DuploTrain),          MP_ROM_PTR(&pb_type_duplo_train)                   },
     #endif
     #if PYBRICKS_PY_PUPDEVICES_MARIO
     { MP_ROM_QSTR(MP_QSTR_Mario),               MP_ROM_PTR(&pb_type_remote)                        },